fix: validate path of `set` to protect against prototype pollution and improve setting to not trigger setters
