+++ .changeset/moody-lizards-taste.md
@@ -1,0 +1,5 @@
+---
+"@elgato/utils": patch
+---
+
+Updated `set` to error when an unsafe path is detected to prevent prototype pollution.
