+++ packages/utils/src/index.ts
@@ -3,7 +3,9 @@ export * from "./event-emitter.js";
 export * from "./explicit-resource-management/index.js";
 export * from "./json.js";
 export * from "./lazy.js";
-export * from "./objects.js";
+export { freeze } from "./objects/freeze.js";
+export { get } from "./objects/get.js";
+export { set } from "./objects/set.js";
 export * from "./parsers.js";
 export * from "./processes.js";
 export * from "./promises.js";
