+++ packages/utils/src/objects/freeze.ts
@@ -1,0 +1,10 @@
+/**
+ * Prevents the modification of existing property attributes and values on the value, and all of its child properties, and prevents the addition of new properties.
+ * @param value Value to freeze.
+ */
+export function freeze<T>(value: T): void {
+	if (value !== undefined && value !== null && typeof value === "object" && !Object.isFrozen(value)) {
+		Object.freeze(value);
+		Object.values(value).forEach(freeze);
+	}
+}
